Fix logic errors in zfs setup

+++ packages/setup-zfs-machine/setup-host-with-zfs.sh
@@ -41,38 +41,56 @@ echo "Got \`$(gum style --foreground ${BLUE} "TARGET_DEVICE")=$(gum style --fore
 sgdisk -O $TARGET_DEVICE
 
 gum style "
-This will irrevocably destroy all data on \`TARGET_DEVICE=${TARGET_DEVICE}\`!!!
+Do you want to destroy the partitions or use them 'as is'? 
+If you plan to use them as is, there should be the following
+partitions and zpools on the device:
 
-A FAT32 EFI system partition will be created as the first partition.
+Partition 1 should be a FAT32 partition that is EFI (ef00).
 
-An zfs partition will be created as the second partition."
+Partion 2 is used for the Zpool and should already have these datasets:
 
-gum confirm "Are you ready to setup this host (including destroying all data on the disk)?" || (echo "Ok!" && exit 1)
+Zpool Arrangement:
+  NIXROOT/root
+  NIXROOT/home
+  NIXROOT/nix
+  NIXROOT/reserved
 
+"
+DESTROY=$(gum choose "DESTROY" "KEEP")
+if [ "$DESTROY" == "DESTROY" ]; then
 
-gum style --bold --foreground "${RED}" "Destroying existing partitions on \`TARGET_DEVICE=${TARGET_DEVICE}\` in 10..."
-sleep 1
-gum style --bold --foreground "${RED}" "Formatting"
+  gum style "
+    This will irrevocably destroy all data on \`TARGET_DEVICE=${TARGET_DEVICE}\`!!!
 
-umount -r /mnt || true
-umount -r "${TARGET_DEVICE}" || true
-sgdisk --zap-all "${TARGET_DEVICE}"
-sgdisk -o "${TARGET_DEVICE}"
-partprobe || true
+    A FAT32 EFI system partition will be created as the first partition.
 
-sgdisk -n 0:0:+10G -t 0:ef00 -c 0:efi "${TARGET_DEVICE}"
+    An zfs partition will be created as the second partition."
 
-mkfs.fat -F 32 "${TARGET_DEVICE}"1
-fatlabel "${TARGET_DEVICE}"1 NIXBOOT
+  gum confirm "Are you ready to setup this host (including destroying all data on the disk)?" || (echo "Ok!" && exit 1)
+  gum style --bold --foreground "${RED}" "Destroying existing partitions on \`TARGET_DEVICE=${TARGET_DEVICE}\` in 10..."
+ 
+  sleep 1
+  gum style --bold --foreground "${RED}" "Formatting"
 
-# Create a partition for / using the remaining space
-sgdisk -n 2:0:0 -t 0:8300 -c 0:NIXROOT "${TARGET_DEVICE}" 
+  umount -r /mnt || true
+  umount -r "${TARGET_DEVICE}" || true
+  sgdisk --zap-all "${TARGET_DEVICE}"
+  sgdisk -o "${TARGET_DEVICE}"
+  partprobe || true
 
-echo "Do you want to encrypt your disk? Typically you should only say no here if you plan to restart the host remotely / without user interaction."
-ENCRYPT=$(gum choose "YES" "NO")
-if [ "$ENCRYPT" == "YES" ]; then
-  # Create an encrypted zpool
-  zpool create -f \
+  sgdisk -n 0:0:+10G -t 0:ef00 -c 0:efi "${TARGET_DEVICE}"
+
+  mkfs.fat -F 32 "${TARGET_DEVICE}"1
+  fatlabel "${TARGET_DEVICE}"1 NIXBOOT
+
+  # Create a partition for / using the remaining space
+  sgdisk -n 2:0:0 -t 0:8300 -c 0:NIXROOT "${TARGET_DEVICE}" 
+
+  echo "Do you want to encrypt your disk? Typically you should only say no here if you plan to restart the host remotely / without user interaction."
+  ENCRYPT=$(gum choose "YES" "NO")
+  if [ "$ENCRYPT" == "YES" ]; then
+    # Create an encrypted zpool
+    zpool create -f \
         -o altroot="/mnt" \
  	-o ashift=12 \
 	-o autotrim=on \
@@ -87,9 +105,9 @@ if [ "$ENCRYPT" == "YES" ]; then
 	-O keyformat=passphrase \
 	-O mountpoint=none \
 	NIXROOT "${TARGET_DEVICE}"2
-else
-  # Create an non encrypted zpool
-  zpool create -f \
+  else
+    # Create an non encrypted zpool
+    zpool create -f \
         -o altroot="/mnt" \
  	-o ashift=12 \
 	-o autotrim=on \
@@ -101,14 +119,28 @@ else
 	-O dnodesize=auto \
 	-O mountpoint=none \
 	NIXROOT "${TARGET_DEVICE}"2
+  fi
+  # legacy mount points do not get auto mounted at boot
+  # rather they must be mounted using fstab
+  zfs create -o mountpoint=legacy NIXROOT/root
+  zfs create -o mountpoint=legacy NIXROOT/home
+  zfs create -o mountpoint=legacy NIXROOT/nix
+  # reserved to cope with running out of disk space
+  zfs create -o refreservation=1G -o mountpoint=none NIXROOT/reserved
+
 fi
-# legacy mount points do not get auto mounted at boot
-# rather they must be mounted using fstab
-zfs create -o mountpoint=legacy NIXROOT/root
-zfs create -o mountpoint=legacy NIXROOT/home
-zfs create -o mountpoint=legacy NIXROOT/nix
-# reserved to cope with running out of disk space
-zfs create -o refreservation=1G -o mountpoint=none NIXROOT/reserved
+
+gum style \
+	--foreground 212 --border-foreground 212 --border double \
+	--align center --width 50 --margin "1 2" --padding "2 4" \
+	'Mounting' 'Mounting the partitions and ZFS datasets:
+
+/mnt/boot
+NIXROOT/root /mnt
+NIXROOT/home /mnt/home
+NIXROOT/nix  /mnt/nix
+'
+
 mount -t zfs NIXROOT/root /mnt
 mkdir /mnt/boot
 mkdir /mnt/home
@@ -117,11 +149,17 @@ mkdir /mnt/nix
 mount "${TARGET_DEVICE}"1 /mnt/boot
 mount -t zfs NIXROOT/home /mnt/home
 mount -t zfs NIXROOT/nix /mnt/nix
-nixos-generate-config --root /mnt
+
+hostname ${HOSTNAME}
+
+gum style \
+	--foreground 212 --border-foreground 212 --border double \
+	--align center --width 50 --margin "1 2" --padding "2 4" \
+	'Mounts' $(sudo zfs list;sudo mount | grep NIXROOT;sudo mount | grep boot)
 
 echo "Are you installing an existing flake profile?"
 FLAKE=$(gum choose "YES" "NO")
-if [ "$VM" == "YES" ]; then
+if [ "$FLAKE" == "YES" ]; then
   git clone https://github.com/timlinux/nix-config.git
   cd nix-config
   cd flakes
@@ -130,11 +168,6 @@ if [ "$VM" == "YES" ]; then
   cd ..
   mv /mnt/etc/nixos /mnt/etc/nixos_org
   mv nix-config /mnt/etc/nixos
-  hostname ${HOSTNAME}
-  # Just for info...
-  sudo zfs list
-  sudo mount | grep NIXROOT
-  sudo mount | grep boot
   gum style "
 Partitioning is complete. If you do not have a host entry for this host
 already configured, you need to add it to the hosts folder
@@ -149,6 +182,17 @@ fi
 ##
 ## Remaining steps are of if you are not using flake based install
 ##
+gum style \
+	--foreground 212 --border-foreground 212 --border double \
+	--align center --width 50 --margin "1 2" --padding "2 4" \
+	'Mounting' 'Mounting the partitions and ZFS datasets:
+
+/mnt/boot
+NIXROOT/root /mnt
+NIXROOT/home /mnt/home
+NIXROOT/nix  /mnt/nix
+'
+nixos-generate-config --root /mnt
 
 echo "Are you installing in a VM?"
 VM=$(gum choose "YES" "NO")
